requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=0.19.0

# Bybit API
pybit>=5.0.0
ccxt>=3.0.0

# Technical Analysis
ta-lib>=0.4.25
ta>=0.10.2

# Machine Learning
scikit-learn>=1.1.0
tensorflow>=2.10.0
torch>=1.12.0
xgboost>=1.6.0
lightgbm>=3.3.0
optuna>=3.0.0

# Data Processing
scipy>=1.9.0
scikit-optimize>=0.9.0

# Database & Storage
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.0.0
timescaledb>=0.1.0
parquet>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0

# Backtesting
vectorbt>=0.25.0
backtrader>=1.9.76

# Monitoring & Alerts
prometheus-client>=0.15.0
python-telegram-bot>=20.0
grafana-api>=1.0.3

# MLflow
mlflow>=2.0.0

# Utilities
schedule>=1.2.0
loguru>=0.6.0
websocket-client>=1.4.0
asyncio-mqtt>=0.11.0
click>=8.0.0
rich>=12.0.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.8.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.950
jupyter>=1.0.0

# Infrastructure
docker>=6.0.0
docker-compose>=1.29.0

# Security
cryptography>=3.4.8
keyring>=23.0.0

# Advanced Trading Strategies
scipy>=1.9.0
scikit-learn>=1.1.0
cvxpy>=1.3.0
pulp>=2.6.0

# Reinforcement Learning
gym>=0.21.0
stable-baselines3>=1.8.0
tensorboard>=2.10.0

# Natural Language Processing
transformers>=4.20.0
torch>=1.12.0
nltk>=3.7
textblob>=0.17.1

# News and Sentiment Analysis
newsapi-python>=0.2.6
tweepy>=4.12.0
vaderSentiment>=3.3.2

# On-chain Analysis
web3>=6.0.0
requests>=2.28.0

# Portfolio Optimization
PyPortfolioOpt>=1.5.0
riskfolio-lib>=4.0.0

# Advanced Statistics
statsmodels>=0.13.0
arch>=5.3.0

# Alternative Data Sources
yfinance>=0.1.87
alpha-vantage>=2.3.1
fredapi>=0.5.0
